+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `JSONConfig`: changed to use system default buffer size
+- `JSONSerializer`: added serialization of `java.sql.Date`, `java.sql.Time`, `java.sql.Timestamp`
+- `JSONDeserializer`: added deserialization of `java.sql.Date`, `java.sql.Time`, `java.sql.Timestamp`
+
+### Added
+- `JSONFun`: JSON helper functions
+- `JSONFunTest`: tests for JSON helper functions
 
 ## [0.7] - 2019-07-06
 ### Changed